refactor: split p2p::test_utils into more files
